Rename helm chart directory to Crossplane
We name the chart as "crossplane". Eventually it will be in
upbound-stable/crossplane helm repo.

Crossplane helm chart uses chart name to template
some names / values in the manifests, we were using nameOverride
in UXP v1, but trying to avoid in case some upstream manifest forgot
to incorporate that and to minimize the divergence from upstream.

Signed-off-by: Hasan Turken <turkenh@gmail.com>