chore: disable helm docs generation
The build submodule installs the tool using Go, but we cannot import
golang.mk since it sets up all the Go-related machinery, and this
repository lacks a Go module to handle, causing it to fail. We were not
using it previously either, as we did not import helm.mk, so even when
it was set to true, it had no effect, therefore this change is a no-op.

Signed-off-by: Philippe Scorsolini <p.scorsolini@gmail.com>